ai_engine/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
httpx-auth-awssigv4>=0.1.3
boto3>=1.34.0
pydantic>=2.10.0
pyyaml>=6.0
python-multipart>=0.0.12
